analyzer/requirement.txt
zerorpc
pdfminer==20110515
python-magic
chardet
